requirements.txt
asgiref==3.11.1
Django==4.2.30
sqlparse==0.5.5
typing_extensions==4.15.0
psycopg[binary]>=3.2
uvicorn[standard]>=0.30
daphne>=4.1
channels>=4.1
channels-redis>=4.2
celery[redis]>=5.4
redis>=5.0
hiredis==3.1.0
django-celery-results==2.5.1
django-celery-beat>=2.7
djangorestframework>=3.15
django-filter>=24.2
Pillow>=10.4
django-cors-headers==4.6.0
cryptography==43.0.3
python-dotenv>=1.0
whitenoise>=6.7
gunicorn>=22.0
flower>=2.0

httpx==0.28.1
tenacity==9.0.0
python-magic==0.4.27

python-slugify==8.0.4
python-dateutil==2.9.0
shortuuid==1.0.13
bleach==6.2.0
markdown==3.7
humanize==4.11.0

sentry-sdk[django]==2.19.2
django-health-check==3.18.3
structlog==24.4.0

pytest==8.3.4
pytest-django==4.9.0
pytest-cov==6.0.0
factory-boy==3.3.1
faker==33.1.0
model-bakery==1.20.0
freezegun==1.5.1
responses==0.25.3

ruff==0.8.4
pre-commit==4.0.1
mypy==1.14.0
django-stubs==5.1.1
django-crispy-forms==1.8.0
django-csp==4.0
django-filter==25.1
django-import-export==4.4.0
django-mathfilters==1.0.0
django-model-utils==5.0.0
django-money==3.5.4

python-decouple==3.8


